Added PermissibleTrait to remove boilerplate
