+++ src/pocketmine/permission/PermissibleTrait.php
@@ -1,0 +1,73 @@
+<?php
+
+/*
+ *
+ *  ____            _        _   __  __ _                  __  __ ____
+ * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \
+ * | |_) / _ \ / __| |/ / _ \ __| |\/| | | '_ \ / _ \_____| |\/| | |_) |
+ * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/
+ * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_|
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * @author PocketMine Team
+ * @link http://www.pocketmine.net/
+ *
+ *
+*/
+
+declare(strict_types=1);
+
+namespace pocketmine\permission;
+
+use pocketmine\plugin\Plugin;
+
+trait PermissibleTrait{
+	/** @var PermissibleBase */
+	private $perm;
+
+	protected function initPermissible() : void{
+		assert($this instanceof Permissible);
+		$this->perm = new PermissibleBase($this);
+	}
+
+	protected function destroyPermissible() : void{
+		$this->perm->clearPermissions();
+		$this->perm = null;
+	}
+
+	public function isPermissionSet($name) : bool{
+		return $this->perm->isPermissionSet($name);
+	}
+
+	public function hasPermission($name) : bool{
+		return $this->perm->hasPermission($name);
+	}
+
+	public function addAttachment(Plugin $plugin, string $name = null, bool $value = null) : PermissionAttachment{
+		return $this->perm->addAttachment($plugin, $name, $value);
+	}
+
+	public function removeAttachment(PermissionAttachment $attachment) : void{
+		$this->perm->removeAttachment($attachment);
+	}
+
+	public function recalculatePermissions() : void{
+		$this->perm->recalculatePermissions();
+	}
+
+	public function getEffectivePermissions() : array{
+		return $this->perm->getEffectivePermissions();
+	}
+
+	public function isOp() : bool{
+		return $this->perm->isOp();
+	}
+
+	public function setOp(bool $value) : void{
+		$this->perm->setOp($value);
+	}
+}
